fix(helper_methods): use shallow copy
We want a copy of the object to add other inherited member_data_types_
to. If we use `=`, we get a reference to `self.member_data_types_`
object and end up modifying it, which we do not want to do. Not only
does that change the nml.py representation, it also means that all the
"expanded" lists of `member_data_types_` is held in memory for every
neuroml object which greatly increases the memory foot print of models.

So, calculate the complete list of member_data_types_ when required and
then forget about it, which will allow memory to be garbage collected
and freed up.