fix: py2 build
Use BytesIO for py2, but StringIO for py3.

I'm not entirely sure why this is needed here. It isn't necessarily an
issue in nml.py---it could be in the loaders where we're not being too
careful about py2/py3. The versions of pytables in use are also
different because pytables dropped support for py2 in 3.6.x.