fix: correctly fix py2/3 StringIO error
In python 2, we must use `from StringIO import StringIO` for strings.
This is not the same as `from io import StringIO` in Python2 (or in
Python3).

The former handles strings, the latter handles unicode. For Py2, since
we're not prefixing all strings with `u""` to make them unicode, we need
to use the first one.

So, different `StringIO` classes need to be imported for Python2 and
Python3. The same `from io import StringIO` cannot be shared.

- https://docs.python.org/2.7/library/stringio.html?highlight=stringio#StringIO.StringIO
- https://docs.python.org/2.7/library/io.html?highlight=io%20stringio#io.StringIO