feat(nml.py)!: `group` -> `group_id` for clarity
BREAKING CHANGE: argument of the cell helper functions has changed from
`group` to `group_id` to make it clear to users that it is the id of the
group that is required there, not the object.

The cell builder utilities have not yet been released, so while this is
a breaking change, this will not affect users unless they've been using
the development branch.